+++ cli/js/runtime.ts
@@ -1,6 +1,7 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { core } from "./core.ts";
-import * as dispatch from "./dispatch.ts";
+import * as dispatchMinimal from "./ops/dispatch_minimal.ts";
+import * as dispatchJson from "./ops/dispatch_json.ts";
 import { assert } from "./util.ts";
 import * as util from "./util.ts";
 import { setBuildInfo } from "./build.ts";
@@ -11,12 +12,22 @@ import { Start, start as startOp } from "./ops/runtime.ts";
 
 export let OPS_CACHE: { [name: string]: number };
 
+function getAsyncHandler(opName: string): (msg: Uint8Array) => void {
+  switch (opName) {
+    case "op_write":
+    case "op_read":
+      return dispatchMinimal.asyncMsgFromRust;
+    default:
+      return dispatchJson.asyncMsgFromRust;
+  }
+}
+
 // TODO(bartlomieju): temporary solution, must be fixed when moving
 // dispatches to separate crates
 export function initOps(): void {
   OPS_CACHE = core.ops();
   for (const [name, opId] of Object.entries(OPS_CACHE)) {
-    core.setAsyncHandler(opId, dispatch.getAsyncHandler(name));
+    core.setAsyncHandler(opId, getAsyncHandler(name));
   }
 }
