--- cli/js/types.ts
@@ -1,2 +1,0 @@
-// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-export type TypedArray = Uint8Array | Float32Array | Int32Array;
